Move dashboard logic from cmd to pkg.
This commit moves the dashboard method from prmonitor/cmd/prmonitor
to just prmonitor so that it can be placed under better test coverage.
My next step is some refactoring to make it testable, then modifying
it to fit some new features.
